refactor: move new functions away from init code
